재고량 감소 함수 리팩토링 - #9
