+++ init.py
@@ -2,45 +2,7 @@ import pandas as pd
 import random
 import datetime
 
-from generate_special_event import generate_special_event
-
-
-# 재고량 감소 함수
-def decrease_stock_init(
-    sales_data: list, remaining_stock: int, timestamp: datetime, item_id: int
-) -> tuple:
-    def add_sale(sales_qty: int):
-        nonlocal remaining_stock
-        # 재고 음수 방지
-        remaining_stock = max(remaining_stock - sales_qty, 0)
-        sale_row = {
-            "item_id": item_id,
-            "remaining_stock": remaining_stock,
-            "timestamp": timestamp.strftime("%Y-%m-%d %H:%M:%S"),
-        }
-        sales_data.append(sale_row)
-
-    hour = timestamp.hour
-    prob, sale_range = None, None
-
-    # 새벽 시간인 경우, 30% 확률로 1~3개 판매, 70% 확률로 판매 없음
-    if hour < 6:
-        prob, sale_range = 0.3, (1, 3)
-    # 06 ~ 09시 사이인 경우, 50% 확률로 1~5개 판매, 50% 확률로 판매 없음
-    elif 6 <= hour < 9:
-        prob, sale_range = 0.5, (1, 5)
-    else:
-        prob, sale_range = 1.0, (1, 5)
-
-    if random.random() < prob:
-        sales_qty = random.randint(*sale_range)
-
-        remaining_stock = generate_special_event(
-            hour, remaining_stock
-        )  # 일반적이지 않은 재고 감소 이벤트
-        add_sale(sales_qty)
-
-    return sales_data, remaining_stock
+from decrease_stock import decrease_stock
 
 
 # 임의의 초기 데이터프레임 생성
@@ -66,7 +28,7 @@ def init() -> pd.DataFrame:
             # 간격을 최소 30분으로 두고 이벤트 생성
             minutes_offset = random.randint(30, 300)
             timestamp_sales += datetime.timedelta(minutes=minutes_offset)
-            sales_data, remaining_stock = decrease_stock_init(
+            sales_data, remaining_stock = decrease_stock(
                 sales_data=sales_data,
                 remaining_stock=remaining_stock,
                 timestamp=timestamp_sales,
